+++ dca/src/tests/mod.rs
@@ -1,3 +1,3 @@
 pub mod mock;
-pub mod on_block_initialize;
+pub mod on_initialize;
 pub mod schedule;
